notes-extensions-updater/pyproject.toml
[tool.poetry]
name = "notes-extensions-updater"
version = "0.1.0"
description = ""
authors = ["Eugene Serdyuk <16106844+eserdk@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aiofiles = "^0.7.0"
httpx = "^0.19.0"
PyYAML = "^5.4.1"
environs = "^9.3.4"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
